Add TokenReader.
This is intended as a sort of easy way to work with the tokens produced
by GT::Lexer, since otherwise you basically have to iterate over them
and so on.

It's fairly simply to use, so shouldn't pose too much of a problem for
anyone using it or subclassing it.