Allow caches to be given an existing array of objects to use.
Caches are still permitted to allocate new objects if they run out of
cached objects, but otherwise they'll pull from the given array.

The cache takes ownership of the array, so it should be considered very
bad behavior to try to modify the array afterward. The upside is that
this allows you to use stuff like Vec3Array from snow-math as the
backing for a cache, meaning you don't end up with as many tiny chunks
of memory being separately allocated for a bunch of vectors or what
have you.